# docs/cie-bff.openapi.yaml
openapi: 3.0.0
info:
  title: cie-bff
  version: '1.0'
  description: API definition for CIE bff
  contact:
    email: cyril.baurand@ch.glory-global.com
    name: Cyril Baurand
servers:
  - url: 'http://localhost:3001'
    description: local machine for debug
paths:

  /users:
    parameters: []

    get:
      summary: get available users
      responses:
        '200':
          $ref: '#/components/responses/GetUsersList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      operationId: get-user
      description: Get user list
      tags:
        - User

    post:
      summary: creates new users
      operationId: post-users-userId
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUsers'
              examples:
                example-1:
                  value:
                    userId: cieuser004
                    userName: tom
                    stores:
                      - name: West Store
                    areas:
                      - name: East Area
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUsers'
            examples:
              example-1:
                value:
                  userId: cieuser004
                  userName: tom
                  stores:
                    - name: West Store
                  areas:
                    - name: East Area
      tags:
        - User
      description: creates new users

    

  '/users/{userId}':
    parameters:
      - $ref: '#/components/parameters/userId'

    get:
      summary: get a user info
      responses:
        '200':
          $ref: '#/components/responses/GetUsers'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      operationId: get-users-userid
      description: get a user information
      tags:
        - User

    put:
      summary: update users
      operationId: put-users-userId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PutUsers'
              examples:
                example-1:
                  value:
                    userName: tom
                    stores:
                      - name: West Stores
                    areas:
                      - name: East Area
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      description: update users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutUsers'
            examples:
              example-1:
                value:
                  userName: tom
                  stores:
                    - name: West Stores
                  areas:
                    - name: East Area
      tags:
        - User

    delete:
      summary: delete users
      operationId: delete-users-userId
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - User
      description: delete users

  '/users/{userId}/restore':
    parameters:
      - $ref: '#/components/parameters/userId'

    put:
      summary: restore users
      operationId: put-users-userId-delete
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - User
      description: restore users

components:
  schemas:
    PostUsers:
      title: PostUsers
      type: object
      description: 'Model of POST /Users'
      x-examples:
        example-1:
          userName: tom
          stores:
            - name: West Store
          areas:
            - name: East Area
      properties:
        userId:
          type: string
          description: new user id
          maxLength: 20
          pattern: '^[a-z0-9]+$'
        userName:
          type: string
          description: user name
          maxLength: 40
          pattern: '^[a-z0-9!@#\\$%&\\*\\(\\)~\\?=;:,\\.\\{\\}\\+\\-`\\|\\/<>]+$'
          minLength: 1
        stores:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: store name
                minLength: 1
            required:
              - name
        areas:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: area name
            required:
              - name
      required:
        - userId
        - userName
        - stores
    PutUsers:
      title: PutUsers
      type: object
      description: 'Model of PUT /users'
      x-examples:
        example-1:
          userName: tom
          stores:
            - name: West Store
          areas:
            - name: East Area
      properties:
        userName:
          type: string
          description: user name
          maxLength: 40
          pattern: '^[a-z0-9!@#\\$%&\\*\\(\\)~\\?=;:,\\.\\{\\}\\+\\-`\\|\\/<>]+$'
          minLength: 1
        stores:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: store name
                minLength: 1
            required:
              - name
        areas:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: area name
            required:
              - name
      required:
        - userName
        - stores
  securitySchemes: {}

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 400
                message: Invalid Request Body
            example-2:
              value:
                code: 400
                message: Invalid Request Path Params
            example-3:
              value:
                code: 400
                message: Invalid Request Params
    Unauthorised:
      description: Unauthorised
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 401
                message: Unauthorized
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 403
                message: Forbidden
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 404
                message: Not Found
    ConflictError:
      description: Conflict Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 409
                message: Conflict Error
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
              message:
                type: string
          examples:
            example-1:
              value:
                code: 500
                message: Unexpected Error
    
    GetUsersList:
      description: |-
        Response of Get /users
        If the number of results obtained is zero, status returns 200.
      content:
        application/json:
          schema:
            type: object
            properties:
              userId:
                type: string
                description: user ID(User ID of user)
              userName:
                type: string
                description: user name
              deleteFlag:
                type: boolean
                description: delete flag(true is deleted)
              areas:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: area name
                  required:
                    - name
              stores:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: store name
                  required:
                    - name
            required:
              - userId
              - userName
              - deleteFlag
              - areas
              - stores
          examples:
            example-1:
              value:
                - userId: cisuser1
                  userName: cis-user-1-1
                  deleteFlag: false
                  areas:
                    - name: East Area
                  stores:
                    - name: Himeji HQ
                - userId: cisuser2
                  userName: cis-user-1-2
                  deleteFlag: false
                  areas:
                    - name: East Area
                    - name: West Area
                    - name: South Area
                  stores:
                    - name: Himeji HQ
                    - name: Tokyo HQ
                    - name: Shinagawa Office
                    - name: Saitama Factory
            example-2:
              value:
                []
      headers: {}

    GetUsers:
      description: |-
        Response of Get /users/{userId}
        If the number of results obtained is zero, status returns 200.
      content:
        application/json:
          schema:
            type: object
            properties:
              userId:
                type: string
                description: user ID(User ID of user)
              userName:
                type: string
                description: user name
              stores:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: store name
                  required:
                    - name
            required:
              - userId
              - userName
              - stores
          examples:
            example-1:
              value:
                userId: cisuser1
                userName: cis-user-1-1
                deleteFlag: false
                areas:
                  - name: East Area
                stores:
                  - name: Himeji HQ
            example-2:
              value:

  parameters:
    areaName:
      name: areaName
      in: query
      required: false
      schema:
        type: string
        minLength: 1
      description: Area name
    storeName:
      name: storeName
      in: query
      required: false
      schema:
        type: string
        minLength: 1
      description: Store name
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        maxLength: 20
        pattern: '^[a-z0-9]+$'
        minLength: 1
      description: User ID

tags:
  - name: User